Improve HTTP/2 scaling performance (#35694)
* Rearrange Http2Connection.SendHeadersAsync to increase scale

Many concurrent streams are currently bottlenecking in SendHeadersAsync.  They:
- take the headers serialization lock
- request a stream credit
- serialize all their headers to a buffer stored on the connection
- take the write lock
- write out the headers
- release the write lock
- release the headers serialization lock

Instead of using a header buffer pooled on the connection, we can instead temporarily grab ArrayPool buffers into which we serialize the headers, which allows us to eliminate the headers serialization lock.  With that, we get:
- request a stream credit
- serialize the headers to a pooled buffer
- take the write lock
- write out the headers
- release the write lock

This has a significant impact on the ability for many concurrent streams to scale, as all of the header serialization work happens outside of the lock.

* Use ContinueWith instead of async/await in Ignore/LogExceptions

The current implementation will always invoke the continuation, even in the majority case where there's no failure (and if there is a failure, it's cheaper not to throw it again).  We can avoid that and decrease both overhead and allocation in the common case.

* Remove CopyToAsync wrapping task from SendRequestBodyAsync

* Reduce contention on semaphore locks

We're seeing a lot of contention trying to acquire the monitors inside of semaphore slims, even when the contention is to release the semaphore, which should entail minimal contention (and delay of a release will just cause more contention).  We can use a small interlocked gate to add a fast path.

* Stop clearing ArrayBuffer's pooled byte arrays

We already weren't clearing as the array grew, nor are we clearing in most other places in the library.  It's not clear why we were clearing in this one spot, but the zero'ing is showing up meaningfully in profiles.

* Add several gRPC known headers

* Avoid dictionary lookup in HttpMethod.Normalize

* Avoid allocating strings for known Content-Types

* Combine AcquireWriteLockAsync into StartWriteAsync

There's not a good reason to keep them separate, and StartWriteAsync is AcquireWriteLockAsync's only caller.

* Remove GetCancelableWaiterTask async method

We can achieve the same thing without the extra async method by putting the cleanup logic into the awaiter's GetResult.

* Address PR feedback